fix for binary reordering format on 64-bit platforms. thanks to Jaakko Väyrynen for identifying the problem
git-svn-id: https://mosesdecoder.svn.sourceforge.net/svnroot/mosesdecoder/trunk/moses@1423 1f5c12ca-751b-0410-a591-d2e778427230

+++ src/PrefixTreeMap.cpp
@@ -3,14 +3,14 @@
 void GenericCandidate::readBin(FILE* f){
   m_PhraseList.clear();
   m_ScoreList.clear();
-  unsigned int num_phrases;
+  uint32_t num_phrases;  // on older compilers, <stdint.h> may need to be included
   fRead(f, num_phrases);
   for(unsigned int i = 0; i < num_phrases; ++i){
 	IPhrase phrase;
 	fReadVector(f, phrase);
 	m_PhraseList.push_back(phrase);
   };
-  unsigned int num_scores;
+  uint32_t num_scores;
   fRead(f, num_scores);
   for(unsigned int j = 0; j < num_scores; ++j){
 	std::vector<float> score;
@@ -20,11 +20,12 @@ void GenericCandidate::readBin(FILE* f){
 };
 
 void GenericCandidate::writeBin(FILE* f) const {
-  fWrite(f, m_PhraseList.size());
+  // cast is necessary to ensure compatibility between 32- and 64-bit platforms
+  fWrite(f, static_cast<uint32_t>(m_PhraseList.size()));
   for(size_t i = 0; i < m_PhraseList.size(); ++i){
 	fWriteVector(f, m_PhraseList[i]);
   }
-  fWrite(f, m_ScoreList.size());
+  fWrite(f, static_cast<uint32_t>(m_ScoreList.size()));
   for(size_t j = 0; j < m_ScoreList.size(); ++j){
 	fWriteVector(f, m_ScoreList[j]);
   }
@@ -32,7 +33,7 @@ void GenericCandidate::writeBin(FILE* f) const {
 
 
 void Candidates::writeBin(FILE* f) const {
-  unsigned s = this->size();
+  uint32_t s = this->size();
   fWrite(f,s);
   for(size_t i = 0; i < s; ++i) {
 	MyBase::operator[](i).writeBin(f);
@@ -40,7 +41,7 @@ void Candidates::writeBin(FILE* f) const {
 }
 
 void Candidates::readBin(FILE* f) {
-  unsigned s;
+  uint32_t s;
   fRead(f,s);
   this->resize(s);
   for(size_t i = 0; i<s; ++i) {
